+++ packages/brick_offline_first/lib/offline_first.dart
@@ -2,3 +2,4 @@ export 'package:brick_core/query.dart';
 export 'package:brick_offline_first/src/offline_first_adapter.dart';
 export 'package:brick_offline_first/src/offline_first_repository.dart';
 export 'package:brick_offline_first/src/offline_first_exception.dart';
+export 'package:brick_offline_first/src/offline_first_policy.dart';