+++ packages/brick_offline_first/test/offline_first/offline_first_repository_test.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
@@ -8,12 +10,44 @@ void main() {
   sqfliteFfiInit();
 
   group('OfflineFirstRepository', () {
-    setUpAll(() async {
-      TestRepository.configure(
-        sqliteDictionary: sqliteModelDictionary,
-      );
+    setUp(() async {
+      TestRepository.configure();
 
       await TestRepository().initialize();
+    });
+
+    tearDown(() {
+      TestRepository.throwOnNextRemoteMutation = false;
+    });
+
+    test('#applyPolicyToQuery', () async {
+      const policy = OfflineFirstGetPolicy.localOnly;
+      final query = TestRepository().applyPolicyToQuery(Query(), get: policy);
+      expect(query?.providerArgs, {'policy': policy.index});
+    });
+
+    group('#delete', () {
+      test('OfflineFirstDeletePolicy.optimisticLocal', () async {
+        final instance = Mounty(name: 'SqliteName');
+        final upserted = await TestRepository().upsert<Mounty>(instance);
+        expect(await TestRepository().sqliteProvider.get<Mounty>(), hasLength(1));
+
+        TestRepository.throwOnNextRemoteMutation = true;
+        await TestRepository().delete<Mounty>(upserted);
+        expect(await TestRepository().sqliteProvider.get<Mounty>(), isEmpty);
+      });
+
+      test('OfflineFirstDeletePolicy.requireRemote', () async {
+        final instance = Mounty(name: 'SqliteName');
+        final upserted = await TestRepository().upsert<Mounty>(instance);
+
+        TestRepository.throwOnNextRemoteMutation = true;
+        expect(
+          () async => await TestRepository()
+              .delete<Mounty>(upserted, policy: OfflineFirstDeletePolicy.requireRemote),
+          throwsA(const TypeMatcher<SocketException>()),
+        );
+      });
     });
 
     group('#get', () {
@@ -43,22 +77,39 @@ void main() {
 
         expect(findByName.first.name, horse.name);
       });
-    });
 
-    test('#getBatched', () async {
-      final results = await TestRepository().getBatched<Mounty>(requireRemote: false);
-      expect(results.first, isA<Mounty>());
-      expect(results.first.name, 'SqliteName');
+      test('OfflineFirstGetPolicy.localOnly', () async {
+        final results = await TestRepository().get<Horse>(policy: OfflineFirstGetPolicy.localOnly);
+
+        expect(results, isEmpty);
+      });
+
+      test('OfflineFirstGetPolicy.awaitRemoteWhenNoneExist', () async {
+        final results = await TestRepository()
+            .get<Mounty>(policy: OfflineFirstGetPolicy.awaitRemoteWhenNoneExist);
+
+        expect(results, isNotEmpty);
+      });
     });
 
     test('#hydrateSqlite / #get requireRest:true', () async {
-      await TestRepository().get<Mounty>(requireRemote: true);
+      await TestRepository().get<Mounty>(policy: OfflineFirstGetPolicy.awaitRemote);
 
       // verify(TestRepository()
       //     .remoteProvider
       //     .client
       //     .get(Uri.parse('http://0.0.0.0:3000/mounties'), headers: anyNamed('headers')));
     }, skip: 'Client is no longer a Mockito instance');
+
+    test('#getBatched', () async {
+      final instance = Mounty(name: 'SqliteName');
+      await TestRepository().upsert<Mounty>(instance);
+      final results = await TestRepository().getBatched<Mounty>(
+        policy: OfflineFirstGetPolicy.localOnly,
+      );
+      expect(results.first, isA<Mounty>());
+      expect(results.first.name, instance.name);
+    });
 
     test('#storeRestResults', () async {
       final instance = Mounty(name: 'SqliteName');
@@ -68,12 +119,24 @@ void main() {
       expect(results.first.primaryKey, greaterThanOrEqualTo(1));
     });
 
-    test('#upsert', () async {
-      final instance = Mounty(name: 'SqliteName');
-      final results = await TestRepository().upsert<Mounty>(instance);
+    group('#upsert', () {
+      test('OfflineFirstUpsertPolicy.optimisticLocal', () async {
+        final instance = Mounty(name: 'SqliteName');
+        final results = await TestRepository().upsert<Mounty>(instance);
 
-      expect(results.name, 'SqliteName');
-      expect(results.primaryKey, greaterThanOrEqualTo(1));
+        expect(results.name, 'SqliteName');
+        expect(results.primaryKey, greaterThanOrEqualTo(1));
+      });
+
+      test('OfflineFirstUpsertPolicy.requireRemote', () async {
+        final instance = Mounty(name: 'SqliteName');
+        TestRepository.throwOnNextRemoteMutation = true;
+        expect(
+          () async => await TestRepository()
+              .upsert<Mounty>(instance, policy: OfflineFirstUpsertPolicy.requireRemote),
+          throwsA(const TypeMatcher<SocketException>()),
+        );
+      });
     });
 
     test('#reset', () async {