+++ packages/brick_offline_first_with_rest/lib/src/offline_queue/rest_offline_queue_client.dart
@@ -21,6 +21,11 @@ class RestOfflineQueueClient extends http.BaseClient {
 
   final Logger _logger;
 
+  /// Describes the type of policy that came from the request, stringified
+  /// from the [OfflineFirstPolicy] enum. The property will be removed before
+  /// forwarding the request to [_inner].
+  static const policyHeader = 'X-Brick-OfflineFirstPolicy';
+
   RestOfflineQueueClient(
     this._inner,
     this.requestManager, {
@@ -30,8 +35,13 @@ class RestOfflineQueueClient extends http.BaseClient {
 
   @override
   Future<http.StreamedResponse> send(http.BaseRequest request) async {
-    final cacheItem = RestRequestSqliteCache(request as http.Request);
+    final cachePolicy = (request as http.Request).headers.remove(policyHeader);
+    final skipCache = cachePolicy == 'requireRemote';
+    final cacheItem = RestRequestSqliteCache(request);
     _logger.finest('sending: ${cacheItem.toSqlite()}');
+
+    // Process the request immediately and forward any warnings to the caller
+    if (skipCache) return await _inner.send(request);
 
     // "Pull" requests are ignored. See documentation of `RequestSqliteCache#requestIsPush`.
     if (cacheItem.requestIsPush) {