+++ packages/brick_graphql/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+## 0.0.1+4
+
+* Supply `context` as `<String, ContextEntry>` instead of type do to [a limitation in JSON serialization](https://stackoverflow.com/a/70538460)
+
+## 0.0.1+3
+
+* Support supplying `context` in `Query#providerArgs`
+
 ## 0.0.1+2
 
 * Rename `fieldsToRuntimeDefinition` to `fieldsToGraphqlRuntimeDefinition`